+++ src/main/java/lotto/factory/ResultGeneratorFactory.java
@@ -3,7 +3,7 @@ package lotto.factory;
 import java.util.List;
 import lotto.domain.Bonus;
 import lotto.domain.Lotto;
-import lotto.service.ResultGenerator;
+import lotto.service.generator.ResultGenerator;
 
 public class ResultGeneratorFactory {
