+++ src/main/java/lotto/service/generator/ResultGenerator.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.service.generator;
 
 import java.util.ArrayList;
 import java.util.List;
